requirements.txt
streamlit==1.25.0
Pillow==9.5.0
roboflow==1.1.3
qrcode==7.4.2

